requirements.txt
Django==3.2.7
python-dotenv==0.19.0
twilio==7.0.0
gunicorn==20.1.0
django-heroku==0.3.1
fuzzywuzzy==0.18.0
imageio==2.9.0
nftlabs-sdk==0.0.6
Pillow==8.4.0
visvis==1.13.0
pygifsicle==1.0.5